% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_build_installer}
\alias{rdesk_build_installer}
\title{Generate and compile InnoSetup script}
\usage{
rdesk_build_installer(
  stage_root,
  out_dir,
  app_name,
  version,
  publisher,
  website,
  license_file,
  icon_file
)
}
\description{
Generate and compile InnoSetup script
}
\keyword{internal}
