% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_build_stub}
\alias{rdesk_build_stub}
\title{Build the stub launcher exe using g++ from Rtools}
\usage{
rdesk_build_stub(stub_cpp, out_exe, app_name)
}
\description{
Build the stub launcher exe using g++ from Rtools
}
\keyword{internal}
