% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rdesk_close_window}
\alias{rdesk_close_window}
\title{Close the native window process}
\usage{
rdesk_close_window(proc)
}
\arguments{
\item{proc}{The processx process object returned by rdesk_open_window}
}
\description{
Close the native window process
}
\keyword{internal}
