% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_copy_dir}
\alias{rdesk_copy_dir}
\title{Deep copy a directory}
\usage{
rdesk_copy_dir(from, to)
}
\arguments{
\item{from}{Source directory}

\item{to}{Destination directory}
}
\description{
Deep copy a directory
}
\keyword{internal}
