% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{rdesk_create_app}
\alias{rdesk_create_app}
\title{Create a new RDesk application project}
\usage{
rdesk_create_app(app_name, path = ".")
}
\arguments{
\item{app_name}{Name of the application}

\item{path}{Directory where the project will be created}
}
\description{
Create a new RDesk application project
}
