% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_detect_r_home}
\alias{rdesk_detect_r_home}
\title{Auto-detect the current R installation root}
\usage{
rdesk_detect_r_home()
}
\description{
Auto-detect the current R installation root
}
\keyword{internal}
