% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{rdesk_encode_message}
\alias{rdesk_encode_message}
\title{Encode a message into JSON for WebSocket transmission}
\usage{
rdesk_encode_message(type, payload = list())
}
\arguments{
\item{type}{Message type string}

\item{payload}{Data payload (list)}
}
\value{
JSON string
}
\description{
Encode a message into JSON for WebSocket transmission
}
\keyword{internal}
