% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_find_7zip}
\alias{rdesk_find_7zip}
\title{Find 7-Zip executable in common Windows locations}
\usage{
rdesk_find_7zip()
}
\value{
Path to 7z.exe
}
\description{
Find 7-Zip executable in common Windows locations
}
\keyword{internal}
