% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_find_gpp}
\alias{rdesk_find_gpp}
\title{Find g++ compiler in common Rtools locations}
\usage{
rdesk_find_gpp()
}
\value{
Path to g++.exe
}
\description{
Find g++ compiler in common Rtools locations
}
\keyword{internal}
