% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_find_r_dir}
\alias{rdesk_find_r_dir}
\title{Find the R directory within an extracted installation}
\usage{
rdesk_find_r_dir(extracted_root)
}
\arguments{
\item{extracted_root}{Path where R was extracted}
}
\value{
Path to the R root (containing bin/)
}
\description{
Find the R directory within an extracted installation
}
\keyword{internal}
