% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotreload.R
\name{rdesk_hotreload_init}
\alias{rdesk_hotreload_init}
\title{Initialize hot reloading tracking}
\usage{
rdesk_hotreload_init(app_dir)
}
\arguments{
\item{app_dir}{Path to the application root directory.}
}
\value{
A list containing file paths as names and modification times (POSIXct) as values.
}
\description{
Initialize hot reloading tracking
}
\keyword{internal}
