% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_install_packages_to}
\alias{rdesk_install_packages_to}
\title{Install packages into a target library using the bundled R
Downloads binary .zip packages from CRAN and installs offline}
\usage{
rdesk_install_packages_to(pkgs, lib_dir, r_version)
}
\description{
Install packages into a target library using the bundled R
Downloads binary .zip packages from CRAN and installs offline
}
\keyword{internal}
