% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_log}
\alias{rdesk_log}
\title{Log a message to the app's log file}
\usage{
rdesk_log(
  message,
  level = "INFO",
  app_name = Sys.getenv("R_APP_NAME", "RDeskApp")
)
}
\arguments{
\item{message}{Message to log}

\item{level}{Log level ("INFO", "WARN", "ERROR")}

\item{app_name}{Optional app name to determine log file}
}
\description{
Log a message to the app's log file
}
\keyword{internal}
