% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_log_dir}
\alias{rdesk_log_dir}
\title{Resolve the bundled log directory for an app}
\usage{
rdesk_log_dir(app_name = Sys.getenv("R_APP_NAME", "RDeskApp"))
}
\description{
Resolve the bundled log directory for an app
}
\keyword{internal}
