% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{rdesk_make_router}
\alias{rdesk_make_router}
\title{Initialize a message router for bidirectional IPC}
\usage{
rdesk_make_router()
}
\value{
A list of functions: register, dispatch, get_handlers
}
\description{
Initialize a message router for bidirectional IPC
}
\keyword{internal}
