% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rdesk_mime_type}
\alias{rdesk_mime_type}
\title{Get MIME type based on file extension}
\usage{
rdesk_mime_type(path)
}
\arguments{
\item{path}{File path}
}
\value{
MIME type string
}
\description{
Get MIME type based on file extension
}
\keyword{internal}
