% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_parse_hotkey}
\alias{rdesk_parse_hotkey}
\title{Parse a hotkey string into modifiers and virtual key codes}
\usage{
rdesk_parse_hotkey(keys)
}
\arguments{
\item{keys}{String like "Ctrl+Shift+A" or "Alt+F4"}
}
\value{
List with modifiers and vk
}
\description{
Parse a hotkey string into modifiers and virtual key codes
}
\keyword{internal}
