% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_poll_jobs}
\alias{rdesk_poll_jobs}
\title{Poll background jobs}
\usage{
rdesk_poll_jobs()
}
\description{
This is called internally by the main event loop to check if any
background tasks have finished. Handles both mirai and callr backends.
}
\keyword{internal}
