% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_prune_runtime}
\alias{rdesk_prune_runtime}
\title{Prune unnecessary files from the portable R runtime to reduce size}
\usage{
rdesk_prune_runtime(runtime_dir)
}
\arguments{
\item{runtime_dir}{Path to the installed R root}
}
\description{
Prune unnecessary files from the portable R runtime to reduce size
}
\keyword{internal}
