% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rdesk_req_id}
\alias{rdesk_req_id}
\title{Generate a unique request ID for dialog round-trips}
\usage{
rdesk_req_id()
}
\value{
A character string ID
}
\description{
Generate a unique request ID for dialog round-trips
}
\keyword{internal}
