% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_resolve_deps}
\alias{rdesk_resolve_deps}
\title{Resolve package dependency tree (excluding base/recommended packages)}
\usage{
rdesk_resolve_deps(pkgs, avail)
}
\description{
Resolve package dependency tree (excluding base/recommended packages)
}
\keyword{internal}
