% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_resolve_launcher_bin_dir}
\alias{rdesk_resolve_launcher_bin_dir}
\title{Resolve a source-tree launcher binary directory}
\usage{
rdesk_resolve_launcher_bin_dir(project_root)
}
\description{
Resolve a source-tree launcher binary directory
}
\keyword{internal}
