% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_sanitize_log_component}
\alias{rdesk_sanitize_log_component}
\title{Sanitize an app name for filesystem-safe bundled log paths}
\usage{
rdesk_sanitize_log_component(x)
}
\description{
Sanitize an app name for filesystem-safe bundled log paths
}
\keyword{internal}
