% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{rdesk_scaffold_success_msg}
\alias{rdesk_scaffold_success_msg}
\title{Internal success message}
\usage{
rdesk_scaffold_success_msg(name, app_dir, data_source, viz_type, use_async)
}
\description{
Internal success message
}
\keyword{internal}
