% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_start_daemons}
\alias{rdesk_start_daemons}
\title{Start the mirai daemon pool}
\usage{
rdesk_start_daemons()
}
\value{
Invisible number of workers started.
}
\description{
Called once at App$run() startup when rdesk.async_backend is "mirai".
}
\keyword{internal}
