% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{rdesk_start_server}
\alias{rdesk_start_server}
\title{Start the internal httpuv server}
\usage{
rdesk_start_server(port, www_path, ws_handler)
}
\arguments{
\item{port}{TCP port to listen on}

\item{www_path}{Path to the web assets directory}

\item{ws_handler}{Function called when a WebSocket connection is established}
}
\value{
An httpuv server object
}
\description{
Start the internal httpuv server
}
\keyword{internal}
