% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_to_msys_path}
\alias{rdesk_to_msys_path}
\title{Convert a Windows path to MSYS2-style path for Rtools binaries}
\usage{
rdesk_to_msys_path(path)
}
\description{
Convert a Windows path to MSYS2-style path for Rtools binaries
}
\keyword{internal}
