% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_validate_build_inputs}
\alias{rdesk_validate_build_inputs}
\title{Validate build inputs before starting the process}
\usage{
rdesk_validate_build_inputs(app_dir, extra_pkgs, build_installer = FALSE)
}
\description{
Validate build inputs before starting the process
}
\keyword{internal}
