% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_validate_build_inputs_common}
\alias{rdesk_validate_build_inputs_common}
\title{Validate common build inputs across all platforms}
\usage{
rdesk_validate_build_inputs_common(app_dir, extra_pkgs)
}
\arguments{
\item{app_dir}{Path to app directory.}

\item{extra_pkgs}{Character vector of packages.}
}
\description{
Validate common build inputs across all platforms
}
\keyword{internal}
