% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App.R
\name{App}
\alias{App}
\title{RDesk Application Class}
\description{
Create and manage a native desktop window from R.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-App-new}{\code{App$new()}}
\item \href{#method-App-on_ready}{\code{App$on_ready()}}
\item \href{#method-App-on_message}{\code{App$on_message()}}
\item \href{#method-App-send}{\code{App$send()}}
\item \href{#method-App-load_ui}{\code{App$load_ui()}}
\item \href{#method-App-set_menu}{\code{App$set_menu()}}
\item \href{#method-App-dialog_open}{\code{App$dialog_open()}}
\item \href{#method-App-dialog_save}{\code{App$dialog_save()}}
\item \href{#method-App-notify}{\code{App$notify()}}
\item \href{#method-App-set_tray_icon}{\code{App$set_tray_icon()}}
\item \href{#method-App-quit}{\code{App$quit()}}
\item \href{#method-App-run}{\code{App$run()}}
\item \href{#method-App-clone}{\code{App$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-new"></a>}}
\if{latex}{\out{\hypertarget{method-App-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$new(title = "RDesk App", width = 1200, height = 800, icon = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-on_ready"></a>}}
\if{latex}{\out{\hypertarget{method-App-on_ready}{}}}
\subsection{Method \code{on_ready()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$on_ready(fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-on_message"></a>}}
\if{latex}{\out{\hypertarget{method-App-on_message}{}}}
\subsection{Method \code{on_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$on_message(type, fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-send"></a>}}
\if{latex}{\out{\hypertarget{method-App-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$send(type, payload = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-load_ui"></a>}}
\if{latex}{\out{\hypertarget{method-App-load_ui}{}}}
\subsection{Method \code{load_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$load_ui(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-set_menu"></a>}}
\if{latex}{\out{\hypertarget{method-App-set_menu}{}}}
\subsection{Method \code{set_menu()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$set_menu(items)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-dialog_open"></a>}}
\if{latex}{\out{\hypertarget{method-App-dialog_open}{}}}
\subsection{Method \code{dialog_open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$dialog_open(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-dialog_save"></a>}}
\if{latex}{\out{\hypertarget{method-App-dialog_save}{}}}
\subsection{Method \code{dialog_save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$dialog_save(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-notify"></a>}}
\if{latex}{\out{\hypertarget{method-App-notify}{}}}
\subsection{Method \code{notify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$notify(title, msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-set_tray_icon"></a>}}
\if{latex}{\out{\hypertarget{method-App-set_tray_icon}{}}}
\subsection{Method \code{set_tray_icon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$set_tray_icon(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-quit"></a>}}
\if{latex}{\out{\hypertarget{method-App-quit}{}}}
\subsection{Method \code{quit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$quit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-run"></a>}}
\if{latex}{\out{\hypertarget{method-App-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-App-clone"></a>}}
\if{latex}{\out{\hypertarget{method-App-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
