% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_async}
\alias{rdesk_async}
\title{Run a task in the background}
\usage{
rdesk_async(task, args = list(), on_done = NULL, on_error = NULL)
}
\arguments{
\item{task}{A function to run in the background.}

\item{args}{A list of arguments to pass to the task.}

\item{on_done}{Callback function(result) called when the task finishes successfully.}

\item{on_error}{Callback function(error) called if the task fails.}
}
\value{
Invisible job ID.
}
\description{
Run a task in the background
}
