% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_build_linux_app}
\alias{rdesk_build_linux_app}
\title{Build a Linux application bundle}
\usage{
rdesk_build_linux_app(
  app_dir,
  app_name,
  app_version = "1.0.0",
  out_dir = tempdir(),
  runtime_dir = NULL,
  prune_runtime = TRUE,
  portable_r_method = "extract_only",
  build_installer = FALSE,
  use_download = FALSE
)
}
\description{
Build a Linux application bundle
}
\keyword{internal}
