% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_cancel_job}
\alias{rdesk_cancel_job}
\title{Cancel a running background job}
\usage{
rdesk_cancel_job(job_id)
}
\arguments{
\item{job_id}{The ID of the job to cancel.}
}
\value{
Invisible TRUE if cancelled, FALSE if not found.
}
\description{
Cancel a running background job
}
