% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_error_plot}
\alias{rdesk_error_plot}
\title{Generate a base64-encoded error plot}
\usage{
rdesk_error_plot(message = "Error generating plot")
}
\arguments{
\item{message}{Error message to display (optional)}
}
\value{
A base64-encoded PNG string
}
\description{
Generate a base64-encoded error plot
}
