% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_fetch_portable_r}
\alias{rdesk_fetch_portable_r}
\title{Download and extract a portable R installation
Uses the official CRAN Windows binary installer, extracted via 7-Zip}
\usage{
rdesk_fetch_portable_r(r_version, dest_dir)
}
\description{
Download and extract a portable R installation
Uses the official CRAN Windows binary installer, extracted via 7-Zip
}
\keyword{internal}
