% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdesk_is_bundle}
\alias{rdesk_is_bundle}
\title{Check if the app is running in a bundled (standalone) environment}
\usage{
rdesk_is_bundle()
}
\value{
TRUE if running in a bundle, FALSE otherwise
}
\description{
Check if the app is running in a bundled (standalone) environment
}
