% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_jobs_list}
\alias{rdesk_jobs_list}
\title{List currently pending background jobs}
\usage{
rdesk_jobs_list()
}
\value{
A data.frame with job ID, started time, backend, and app ID.
}
\description{
List currently pending background jobs
}
