% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{rdesk_jobs_pending}
\alias{rdesk_jobs_pending}
\title{Check if any background jobs are pending}
\usage{
rdesk_jobs_pending()
}
\value{
Number of pending jobs.
}
\description{
Check if any background jobs are pending
}
