% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{rdesk_make_router}
\alias{rdesk_make_router}
\title{Create an IPC message router}
\usage{
rdesk_make_router()
}
\value{
A list with register() and dispatch() methods
}
\description{
Create an IPC message router
}
\keyword{internal}
