% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\name{rdesk_message}
\alias{rdesk_message}
\title{Construct a standard RDesk IPC message envelope}
\usage{
rdesk_message(
  type,
  payload = list(),
  version = getOption("rdesk.ipc_version", "1.0")
)
}
\arguments{
\item{type}{The message type/action name}

\item{payload}{A list representing the message data}

\item{version}{The contract version (default "1.0")}
}
\value{
A list representing the standard JSON envelope
}
\description{
Construct a standard RDesk IPC message envelope
}
