% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{rdesk_open_window}
\alias{rdesk_open_window}
\title{Open a native window pointing to a URL}
\usage{
rdesk_open_window(
  url,
  title = "RDesk",
  width = 1200,
  height = 800,
  www_path = ""
)
}
\arguments{
\item{url}{The target URL to load}

\item{title}{Window title}

\item{width}{Window width}

\item{height}{Window height}

\item{www_path}{Path to the local assets directory}
}
\value{
A processx process object
}
\description{
Open a native window pointing to a URL
}
\keyword{internal}
