% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App.R
\name{rdesk_service}
\alias{rdesk_service}
\title{Service all active RDesk applications}
\usage{
rdesk_service()
}
\description{
Processes native OS events for all open windows.
Call this periodically if you are running apps with \code{block = FALSE}.
}
