% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{rdesk_validate_build_inputs}
\alias{rdesk_validate_build_inputs}
\title{Validate build inputs before starting the process}
\usage{
rdesk_validate_build_inputs(
  app_dir,
  extra_pkgs,
  build_installer = FALSE,
  portable_r_method = c("extract_only", "installer"),
  runtime_dir = NULL,
  use_download = FALSE
)
}
\arguments{
\item{app_dir}{Path to app directory.}

\item{extra_pkgs}{Character vector of packages.}

\item{build_installer}{Logical.}

\item{portable_r_method}{Method for R portability.}

\item{runtime_dir}{Path to pre-existing runtime.}
}
\description{
Validate build inputs before starting the process
}
\keyword{internal}
